refactor: move pool-decrease components
