+++ packages/web/pages/earn/pool/position/decrease-liquidity.tsx
@@ -1,21 +1,10 @@
-import Footer from "@components/common/footer/Footer";
-import BreadcrumbsContainer from "@containers/breadcrumbs-container/BreadcrumbsContainer";
-import DecreaseLiquidityContainer from "@containers/decrease-liquidity-container/DecreaseLiquidityContainer";
-import HeaderContainer from "@containers/header-container/HeaderContainer";
-import { useLoading } from "@hooks/common/use-loading";
-import { useWindowSize } from "@hooks/common/use-window-size";
-import { useGnotToGnot } from "@hooks/token/use-gnot-wugnot";
-import IncreaseLiquidityLayout from "@views/pool/pool-increase-liquidity/IncreaseLiquidityLayout";
-import { DeviceSize } from "@styles/media";
-import useRouter from "@hooks/common/use-custom-router";
+import { serverSideTranslations } from "next-i18next/serverSideTranslations";
 import { useMemo } from "react";
-import { useGetPoolDetailByPath } from "src/react-query/pools";
+
 import SEOHeader from "@components/common/seo-header/seo-header";
 import { DEFAULT_I18N_NS, SEOInfo } from "@constants/common.constant";
-import { serverSideTranslations } from "next-i18next/serverSideTranslations";
-import { makeRouteUrl } from "@utils/page.utils";
-import { PAGE_PATH, QUERY_PARAMETER } from "@constants/page.constant";
-import { useTranslation } from "react-i18next";
+import useRouter from "@hooks/common/use-custom-router";
+import PoolDecreaseLiquidity from "@views/pool/pool-decrease-liquidity/PoolDecreaseLiquidity";
 
 export async function getStaticProps({ locale }: { locale: string }) {
   return {
@@ -29,33 +18,8 @@ export async function getStaticProps({ locale }: { locale: string }) {
 }
 
 export default function Page() {
-  const { t } = useTranslation();
-  const { width } = useWindowSize();
   const router = useRouter();
-  const poolPath = router.getPoolPath();
   const positionId = router.getPositionId();
-  const { data, isLoading } = useGetPoolDetailByPath(poolPath as string);
-  const { getGnotPath } = useGnotToGnot();
-  const { isLoading: isLoadingCommon } = useLoading();
-
-  const listBreadcrumb = useMemo(() => {
-    return [
-      { title: t("business:pageHeader.earn"), path: "/earn" },
-      {
-        title:
-          width > DeviceSize.mediumWeb
-            ? `${getGnotPath(data?.tokenA).symbol}/${
-                getGnotPath(data?.tokenB).symbol
-              } (${Number(data?.fee) / 10000}%)`
-            : "...",
-        poolPath: router.getPoolPath(),
-        path: makeRouteUrl(PAGE_PATH.POOL, {
-          [QUERY_PARAMETER.POOL_PATH]: poolPath,
-        }),
-      },
-      { title: t("business:pageHeader.decreaseLiqui"), path: "" },
-    ];
-  }, [data, width]);
 
   const seoInfo = useMemo(
     () => SEOInfo["/earn/pool/position/decrease-liquidity"],
@@ -70,17 +34,7 @@ export default function Page() {
         ogTitle={seoInfo?.ogTitle?.()}
         ogDescription={seoInfo?.ogDesc?.()}
       />
-      <IncreaseLiquidityLayout
-        header={<HeaderContainer />}
-        breadcrumbs={
-          <BreadcrumbsContainer
-            listBreadcrumb={listBreadcrumb}
-            isLoading={isLoadingCommon || isLoading}
-          />
-        }
-        increaseLiquidity={<DecreaseLiquidityContainer />}
-        footer={<Footer />}
-      />
+      <PoolDecreaseLiquidity/>
     </>
   );
 }
