--- packages/web/src/containers/my-position-card-list-container/MyPositionCardListContainer.tsx
@@ -1,76 +1,0 @@
-import MyPositionCardList from "@components/common/my-position-card-list/MyPositionCardList";
-import { useWindowSize } from "@hooks/common/use-window-size";
-import React, { useEffect, useState } from "react";
-import { ValuesType } from "utility-types";
-import { useTokenData } from "@hooks/token/use-token-data";
-import useCustomRouter from "@hooks/common/use-custom-router";
-
-export const POSITION_CONTENT_LABEL = {
-  DAILY: "Daily Earnings",
-  BALANCE: "Balance",
-  CLAIMABLE: "Claimable Rewards",
-  VALUE: "Value",
-  APR: "APR",
-  CURRENT_PRICE: "Current Price",
-  MIN_PRICE: "Min Price",
-  MAX_PRICE: "Max Price",
-  STAR_TAG: "✨",
-} as const;
-
-export type POSITION_CONTENT_LABEL = ValuesType<typeof POSITION_CONTENT_LABEL>;
-
-interface MyPositionCardListContainerProps {
-  loadMore?: boolean;
-}
-
-const MyPositionCardListContainer: React.FC<
-  MyPositionCardListContainerProps
-> = () => {
-  const router = useCustomRouter();
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const [currentIndex, setCurrentIndex] = useState(0);
-  const { width } = useWindowSize();
-  const [mobile, setMobile] = useState(false);
-  const handleResize = () => {
-    if (typeof window !== "undefined") {
-      window.innerWidth < 1000 ? setMobile(true) : setMobile(false);
-    }
-  };
-  const { tokenPrices = {} } = useTokenData();
-
-  useEffect(() => {
-    handleResize();
-    window.addEventListener("resize", handleResize);
-    return () => {
-      window.removeEventListener("resize", handleResize);
-    };
-  }, []);
-
-  const movePoolDetail = (id: string) => {
-    router.movePageWithPoolPath("POOL", id);
-  };
-
-  const onClickLoadMore = () => {
-    // Todo
-  };
-
-  return (
-    <MyPositionCardList
-      loadMore={true}
-      isFetched={true}
-      isLoading={false}
-      onClickLoadMore={onClickLoadMore}
-      positions={[]}
-      movePoolDetail={movePoolDetail}
-      currentIndex={currentIndex}
-      mobile={mobile}
-      width={width}
-      showPagination={false}
-      showLoadMore={false}
-      themeKey="dark"
-      tokenPrices={tokenPrices}
-    />
-  );
-};
-
-export default MyPositionCardListContainer;
