+++ packages/web/src/views/pool/pool-decrease-liquidity/components/decrease-liquidity/DecreaseLiquidity.styles.ts
@@ -1,10 +1,10 @@
-import { fonts } from "@constants/font.constant";
-import mixins from "@styles/mixins";
-
-import styled from "@emotion/styled";
-import { Theme } from "@emotion/react";
 import { css } from "@emotion/css";
+import { Theme } from "@emotion/react";
+import styled from "@emotion/styled";
+
+import { fonts } from "@constants/font.constant";
 import { media } from "@styles/media";
+import mixins from "@styles/mixins";
 
 export const DecreaseLiquidityWrapper = styled.div`
   ${mixins.flexbox("column", "flex-start", "flex-start")};
