+++ packages/web/src/views/pool/pool-decrease-liquidity/components/decrease-pool-info/DecreasePoolInfo.styles.tsx
