+++ packages/web/src/views/pool/pool-decrease-liquidity/components/decrease-position-modal/DecreasePositionModal.styles.ts
