+++ packages/web/src/views/pool/pool-decrease-liquidity/components/decrease-position-modal/decrease-max-min/DecreaseMaxMin.styles.ts
