+++ packages/web/src/views/pool/pool-decrease-liquidity/components/decrease-select-position/DecreaseAmount.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { useTranslation } from "react-i18next";
 
 import Range from "@components/common/range/Range";
-import { IPooledTokenInfo } from "@hooks/decrease/use-decrease-handle";
+import { IPooledTokenInfo } from "@views/pool/pool-decrease-liquidity/hooks/use-decrease-handle";
 import { TokenModel } from "@models/token/token-model";
 
 import DecreasePoolInfo from "../decrease-pool-info/DecreasePoolInfo";
