+++ packages/web/src/views/pool/pool-decrease-liquidity/components/decrease-select-position/DecreaseSelectPosition.styles.ts
@@ -1,7 +1,9 @@
 import styled from "@emotion/styled";
+
 import { fonts } from "@constants/font.constant";
 import mixins from "@styles/mixins";
 import { media } from "@styles/media";
+
 import { DecreaseLiquidityBoxStyle } from "../decrease-liquidity/DecreaseLiquidity.styles";
 
 interface Props {
