+++ packages/web/src/views/pool/pool-decrease-liquidity/components/decrease-select-position/DecreaseSelectPosition.tsx
@@ -7,7 +7,7 @@ import MissingLogo from "@components/common/missing-logo/MissingLogo";
 import RangeBadge from "@components/common/range-badge/RangeBadge";
 import { RANGE_STATUS_OPTION } from "@constants/option.constant";
 import { useWindowSize } from "@hooks/common/use-window-size";
-import { IPriceRange } from "@hooks/decrease/use-decrease-handle";
+import { IPriceRange } from "@views/pool/pool-decrease-liquidity/hooks/use-decrease-handle";
 import { TokenModel } from "@models/token/token-model";
 import { DEVICE_TYPE } from "@styles/media";
